Added fiverr_listings_stats processing to FiverrJsonImporter, more...
- Modified currency__rate from integer to float in fiverr_listings, fiverr_gigs, fiverr_seller_profiles
- Added stats_processed_at and stats_processed_status to fiverr_listings
- Added fiverr_listings_stats table migration
- Added stats helper methods to FiverrJsonImporter:
  - getSellerLevelNumeric()
  - calculateSellerLevelStats()
  - extractFacetCount()
  - categorizeGigType()
  - computeListingsStatsRow()

+++ database/migrations/2025_09_06_005341_create_fiverr_gigs_table.php
@@ -231,7 +231,7 @@ return new class () extends Migration {
 
             // Currency fields
             $table->text('currency__name')->nullable();
-            $table->integer('currency__rate')->nullable();
+            $table->float('currency__rate')->nullable();
             $table->text('currency__template')->nullable();
             $table->boolean('currency__forceRound')->nullable();
             $table->integer('currency__forceRoundFromAmount')->nullable();
